Initialize the project
Let it be known that I am trying to make this boss level within 3 days to try to get into a gaming studio.
I currently work at walmart for $15.30 and would like to be paid more and do something more meaningful in life.

This is a 3d version of the game "Square" that I made 3 years ago.