Fix rotation warping by adding independent axis
The rotation of the camera used to be handled by only
the orbit center. Since the local axis of the orbit center
was used, whenever there was a movement of one axis,
all of the axis would be rotated. This would lead to the
z axis being tilted whenever a vertical mouse motion was used.

If a horization axis was used while the z axis was still tilted,
the new rotation would be tilted.

However, by creating an idependent node that handles vertical pivots,
the z axis (the orbit center itself) would be unaffected whenever
there was yaw action.