Implement player rotation towards direction of motion
This was done through a lot of pain and tears.

A new function that is designed to sync the player rotation
to the direciton of motion was added.

It takes the signed angle between the z axis of the player and the
camera and uses that to determine the desired location.