Fix direcitonal issues on bullet
Have the default direction face north
Add a fire() method that will change the velocity
of the bullet.

(When the bullet is instanciated, it's basis is in the default
orientation. When the rotation is directly modified, the velocity vector
when set by _ready() is in the default direciton. fire() allows the
bullet's velocity to be set in its regular direciton.