Fix player bullet orientation issues
The bullet was being set to the transform of the bulletSource which was
the local transform. But when the bullet's transform was being set it
was the bullet's global transform. This resulted in the bullet being
oriented straight with the bullet_source but when those coordinates were
directly applied, the bullet would only shoot along that local transform
which doesn't change locally (as the bullet_source is coupled to the
player).

This is fixed by using global_transforms instead