Have Entity exist in node tree to allow stun
Godot doesn't run the timers unless the are in the scene tree. Entity is
not meant to be directly used as a node object, so it will be added to
the scene tree whenever the game begins so for editing purposes, it
exists as a a pure object within the code. And whenever the game runs,
it exists within the scene tree so the game engine will run the timer.