Create Player Input Handler
The Input Handler is responsible for giving out signals that
signify an input event. As well as supplying relevant input information.

Because it is signal heavy, it will be designed to be placed as an
explicit node for the Player. This design choice matches Godot Node
design philosophy because a similar thing is done with PhysicsBodies and
CollisionShapes.  (A child node that serves to encapsulate a specific function of the parent node